chore: Appease clippy

+++ core/src/avm1/globals/color.rs
@@ -134,7 +134,7 @@ fn set_rgb<'gc>(
         let rgb = args
             .get(0)
             .unwrap_or(&Value::Undefined)
-            .coerce_to_i32(activation)? as i32;
+            .coerce_to_i32(activation)?;
         let [b, g, r, _] = rgb.to_le_bytes();
 
         let mut base = target.base_mut(activation.context.gc_context);
